refactor: move error handler registration to shared activation
