+++ packages/toolkit/src/extensionShared.ts
@@ -9,6 +9,8 @@
  */
 
 import vscode from 'vscode'
+import * as nls from 'vscode-nls'
+
 import globals, { initialize } from './shared/extensionGlobals'
 import { join } from 'path'
 import { Commands } from './shared/vscode/commands2'
@@ -30,6 +32,14 @@ import { initializeAwsCredentialsStatusBarItem } from './auth/ui/statusBarItem'
 import { RegionProvider } from './shared/regions/regionProvider'
 import { ChildProcess } from './shared/utilities/childProcess'
 import { isInBrowser } from './common/browserUtils'
+import { registerErrorHandler as registerCommandErrorHandler } from './shared/vscode/commands2'
+import { ToolkitError, isUserCancelledError, resolveErrorMessageToDisplay } from './shared/errors'
+import { getLogger } from './shared/logger'
+import { showMessageWithUrl } from './shared/utilities/messages'
+import { Logging } from './shared/logger/commands'
+import { registerWebviewErrorHandler } from './webviews/server'
+
+let localize: nls.LocalizeFunc
 
 /**
  * Activation/setup code that is shared by the regular (nodejs) extension AND browser-compatible extension.
@@ -39,6 +49,17 @@ import { isInBrowser } from './common/browserUtils'
  * "shared" function since it breaks in browser. So for now the caller must provide it.
  */
 export async function activateShared(context: vscode.ExtensionContext, getRegionProvider: () => RegionProvider) {
+    localize = nls.loadMessageBundle()
+
+    registerCommandErrorHandler((info, error) => {
+        const defaultMessage = localize('AWS.generic.message.error', 'Failed to run command: {0}', info.id)
+        void logAndShowError(error, info.id, defaultMessage)
+    })
+
+    registerWebviewErrorHandler((error: unknown, webviewId: string, command: string) => {
+        logAndShowWebviewError(error, webviewId, command)
+    })
+
     // Setup the logger
     const toolkitOutputChannel = vscode.window.createOutputChannel('AWS Toolkit', { log: true })
     await activateLogger(context, toolkitOutputChannel)
@@ -107,4 +128,56 @@ async function getMachineId(): Promise<string> {
     }
     const proc = new ChildProcess('hostname', [], { collect: true, logging: 'no' })
     return (await proc.run()).stdout.trim() ?? 'unknown-host'
+}
+
+/**
+ * Logs the error. Then determines what kind of error message should be shown, if
+ * at all.
+ *
+ * @param error The error itself
+ * @param topic The prefix of the error message
+ * @param defaultMessage The message to show if once cannot be resolved from the given error
+ *
+ * SIDE NOTE:
+ * This is only being used for errors from commands and webview, there's plenty of other places
+ * (explorer, nodes, ...) where it could be used. It needs to be apart of some sort of `core`
+ * module that is guaranteed to initialize prior to every other Toolkit component.
+ * Logging and telemetry would fit well within this core module.
+ */
+async function logAndShowError(error: unknown, topic: string, defaultMessage: string) {
+    if (isUserCancelledError(error)) {
+        getLogger().verbose(`${topic}: user cancelled`)
+        return
+    }
+    const logsItem = localize('AWS.generic.message.viewLogs', 'View Logs...')
+    const logId = getLogger().error(`${topic}: %s`, error)
+    const message = resolveErrorMessageToDisplay(error, defaultMessage)
+
+    if (error instanceof ToolkitError && error.documentationUri) {
+        await showMessageWithUrl(message, error.documentationUri, 'View Documentation', 'error')
+    } else {
+        await vscode.window.showErrorMessage(message, logsItem).then(async resp => {
+            if (resp === logsItem) {
+                await Logging.declared.viewLogsAtMessage.execute(logId)
+            }
+        })
+    }
+}
+
+/**
+ * Show a webview related error to the user + button that links to the logged error
+ *
+ * @param err The error that was thrown in the backend
+ * @param webviewId Arbitrary value that identifies which webview had the error
+ * @param command The high level command/function that was run which triggered the error
+ */
+function logAndShowWebviewError(err: unknown, webviewId: string, command: string) {
+    // HACK: The following implementation is a hack, influenced by the implementation of handleError().
+    // The userFacingError message will be seen in the UI, and the detailedError message will provide the
+    // detailed information in the logs.
+    const detailedError = ToolkitError.chain(err, `Webview backend command failed: "${command}()"`)
+    const userFacingError = ToolkitError.chain(detailedError, 'Webview error')
+    logAndShowError(userFacingError, `webviewId="${webviewId}"`, 'Webview error').catch(e => {
+        getLogger().error('logAndShowError failed: %s', (e as Error).message)
+    })
 }
